Static version of demo_split_building is effective.
As for previous commits, what can be shared between static and temporal
remains in PythonCallingDocker/ and the respective specializations is
respectively dispatched in DemoStatic/ and DemoTemporal/ sub-directories.
