+++ Computations/3DTiles/LyonTemporal/PythonCallingDocker/DemoStatic/run_split_buildings_static.py
@@ -1,0 +1,12 @@
+import os
+import sys
+import logging
+import demo_workflow_static as workflow
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+from run_demo_split_buildings import run_demo_split_buildings
+
+
+if __name__ == '__main__': 
+    split = workflow.demo_split
+    run_demo_split_buildings(split)
