+++ Computations/3DTiles/LyonTemporal/PythonCallingDocker/DemoTemporal/demo_workflow_temporal.py
@@ -3,7 +3,7 @@ import logging
 import time
 from demo_lyon_metropole_dowload_and_sanitize_temporal \
   import DemoLyonMetropoleDowloadAndSanitizeTemporal
-from demo_split_buildings import DemoSplitBuildings
+from demo_split_buildings_temporal import DemoSplitBuildingsTemporal
 from demo_strip_attributes import DemoStrip
 from demo_extract_building_dates import DemoExtractBuildingDates
 from demo_load_3dcitydb import DemoLoad3DCityDB
@@ -13,7 +13,7 @@ from demo_3dcitydb_server import Demo3dCityDBServer
 # Definition of the workflow by defining its nodes and connections
 demo_download = DemoLyonMetropoleDowloadAndSanitizeTemporal('BATI', 'stage_1')
 
-demo_split = DemoSplitBuildings()
+demo_split = DemoSplitBuildingsTemporal()
 demo_split.set_results_dir('stage_2') 
 demo_split.set_input_demo(demo_download)
